--- data_discovery_ai/core/agents/trainerAgent.py
@@ -1,2 +1,0 @@
-# task agent for training models
-# TODO: move training related code here
