+++ data_discovery_ai/server.py
@@ -1,5 +1,5 @@
 from fastapi import FastAPI
-from data_discovery_ai.routes import router as api_router
+from data_discovery_ai.core.routes import router as api_router
 
 
 app = FastAPI()
